pyproject.toml
[tool.poetry]
name = "college-genai-bot"
version = "0.1.0"
description = "GenAI bot for college website with LangChain, MongoDB, and Gemini"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"  # Specify compatible Python versions
fastapi = "^0.110.0"
uvicorn = "^0.27.0"
langchain = "0.1.12"  # Pin to specific versions
langchain-community = "0.0.38"
langchain-google-genai = "^0.0.8"
beautifulsoup4 = "^4.12.3"
sqlalchemy = "^2.0.27"
pypdf = "^4.0.1"
pymongo = "^4.6.1"
google-generativeai = "^0.3.2"
pydantic = "^2.6.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"